Break monolithic VoiceIO into subcomponents to improve workflow.  No change in API.
