+++ Sources/VoiceKitCore/PermissionBridge.swift
@@ -1,0 +1,44 @@
+//
+//  PermissionBridge.swift
+//  VoiceKit
+//
+//  Generated by GPT-5 (OpenAI)
+//  collaborator: robert
+//  date: 09-15-2025
+//
+
+import Foundation
+@preconcurrency import Speech
+@preconcurrency import AVFoundation
+#if os(macOS)
+import AVFoundation.AVCaptureDevice
+#endif
+
+enum PermissionBridge {
+
+    nonisolated static func awaitSpeechAuth() async -> SFSpeechRecognizerAuthorizationStatus {
+        await withCheckedContinuation { (c: CheckedContinuation<SFSpeechRecognizerAuthorizationStatus, Never>) in
+            SFSpeechRecognizer.requestAuthorization { status in
+                c.resume(returning: status)
+            }
+        }
+    }
+
+    #if os(iOS)
+    nonisolated static func awaitMicPermission() async -> Bool {
+        await withCheckedContinuation { (c: CheckedContinuation<Bool, Never>) in
+            AVAudioApplication.requestRecordPermission { granted in
+                c.resume(returning: granted)
+            }
+        }
+    }
+    #elseif os(macOS)
+    nonisolated static func awaitMicPermission() async -> Bool {
+        await withCheckedContinuation { (c: CheckedContinuation<Bool, Never>) in
+            AVCaptureDevice.requestAccess(for: .audio) { granted in
+                c.resume(returning: granted)
+            }
+        }
+    }
+    #endif
+}
