+++ Sources/VoiceKitCore/VoicePublic.swift
@@ -1,0 +1,154 @@
+//
+//  VoicePublic.swift
+//  VoiceKit
+//
+//  Generated by GPT-5 (OpenAI)
+//  collaborator: robert
+//  date: 09-15-2025
+//
+
+import Foundation
+import CoreGraphics
+
+// MARK: - Public results and errors
+
+public struct VoiceResult: Sendable {
+    public let transcript: String
+    public let recordingURL: URL?
+    public init(transcript: String, recordingURL: URL?) {
+        self.transcript = transcript
+        self.recordingURL = recordingURL
+    }
+}
+
+public struct SimpleError: LocalizedError, Sendable {
+    public var message: String
+    public init(_ message: String) { self.message = message }
+    public var errorDescription: String? { message }
+}
+
+// MARK: - VoiceIO API (main-actor)
+
+@MainActor
+public protocol VoiceIO: AnyObject {
+    var onListeningChanged: ((Bool) -> Void)? { get set }
+    var onTranscriptChanged: ((String) -> Void)? { get set }
+    var onLevelChanged: ((CGFloat) -> Void)? { get set }
+    var onTTSSpeakingChanged: ((Bool) -> Void)? { get set }
+    var onTTSPulse: ((CGFloat) -> Void)? { get set }
+    var onStatusMessageChanged: ((String?) -> Void)? { get set }
+
+    func ensurePermissions() async throws
+    func configureSessionIfNeeded() async throws
+
+    func speak(_ text: String) async
+    func listen(timeout: TimeInterval, inactivity: TimeInterval, record: Bool) async throws -> VoiceResult
+
+    func prepareBoosted(url: URL, gainDB: Float) async throws
+    func startPreparedBoosted() async throws
+    func playBoosted(url: URL, gainDB: Float) async throws
+
+    func stopAll()
+    func hardReset()
+}
+
+@MainActor
+public extension VoiceIO {
+    func prepareBoosted(url: URL) async throws { try await prepareBoosted(url: url, gainDB: 0) }
+    func playBoosted(url: URL) async throws { try await playBoosted(url: url, gainDB: 0) }
+    func queueSFX(url: URL, gainDB: Float = 0) async throws { try await prepareBoosted(url: url, gainDB: gainDB) }
+
+    // Convenience that passes context when using RealVoiceIO
+    func listen(timeout: TimeInterval,
+                inactivity: TimeInterval,
+                record: Bool,
+                context: RecognitionContext = .init()) async throws -> VoiceResult {
+        if let real = self as? RealVoiceIO {
+            real.setRecognitionContext(context)
+        }
+        return try await listen(timeout: timeout, inactivity: inactivity, record: record)
+    }
+}
+
+// MARK: - TTS models and configuration
+
+public struct TTSVoiceInfo: Identifiable, Hashable, Codable, Sendable {
+    public let id: String
+    public let name: String
+    public let language: String
+    public init(id: String, name: String, language: String) {
+        self.id = id; self.name = name; self.language = language
+    }
+}
+
+public struct TTSVoiceProfile: Identifiable, Hashable, Codable, Sendable {
+    public let id: String
+    public var displayName: String
+    public var rate: Float
+    public var pitch: Float
+    public var volume: Float
+    public var isSelected: Bool
+    public var isHidden: Bool
+    public init(id: String, displayName: String, rate: Float, pitch: Float, volume: Float, isSelected: Bool = false, isHidden: Bool = false) {
+        self.id = id; self.displayName = displayName; self.rate = rate; self.pitch = pitch; self.volume = volume
+        self.isSelected = isSelected; self.isHidden = isHidden
+    }
+}
+
+public struct TTSMasterControl: Codable, Sendable {
+    public var volume: Float
+    public var pitchVariation: Float
+    public var rateVariation: Float
+    public init(volume: Float = 1.0, pitchVariation: Float = 0.0001, rateVariation: Float = 0.0001) {
+        self.volume = volume; self.pitchVariation = pitchVariation; self.rateVariation = rateVariation
+    }
+}
+
+@MainActor
+public protocol TTSConfigurable: AnyObject {
+    func availableVoices() -> [TTSVoiceInfo]
+    func setVoiceProfile(_ profile: TTSVoiceProfile)
+    func getVoiceProfile(id: String) -> TTSVoiceProfile?
+    func setDefaultVoiceProfile(_ profile: TTSVoiceProfile)
+    func getDefaultVoiceProfile() -> TTSVoiceProfile?
+    func setMasterControl(_ master: TTSMasterControl)
+    func getMasterControl() -> TTSMasterControl
+    func speak(_ text: String, using voiceID: String?) async
+    func stopSpeakingNow()
+}
+
+// MARK: - Recognition context
+
+public struct RecognitionContext: Sendable {
+    public enum Expectation: Sendable {
+        case freeform
+        case name(allowed: [String])
+        case number
+    }
+    public var expectation: Expectation
+    public init(expectation: Expectation = .freeform) { self.expectation = expectation }
+}
+
+public extension RecognitionContext {
+    static var numericContextualStrings: [String] {
+        let digits = (0...20).map { String($0) } + ["30","40","50","60","70","80","90","100"]
+        let words = ["zero","one","two","three","four","five","six","seven","eight","nine",
+                     "ten","eleven","twelve","thirteen","fourteen","fifteen",
+                     "sixteen","seventeen","eighteen","nineteen","twenty",
+                     "thirty","forty","fifty","sixty","seventy","eighty","ninety","hundred"]
+        return digits + words
+    }
+}
+
+// MARK: - Operation gate
+
+public actor VoiceOpGate {
+    private var busy = false
+    public init() {}
+    public func acquire() async {
+        while busy { try? await Task.sleep(nanoseconds: 200_000) }
+        busy = true
+    }
+    public func release() async { busy = false }
+    public func forceClear() async { busy = false }
+}
